+++ pages/projects/[id]/index.js
@@ -134,16 +134,20 @@ const QuotationsList = ({ project = {}, inProgress, awarded, completed }) => {
       confirm({
         title: '¿Está seguro de que quiere eliminar el archivo?', //<IntlMessages id="app.common.confirmDeleteFile" />,
         onOk: () => {
-          deleteFile(
-            file.url.split('files/')[1].split('/')[0],
-            data => {
-              successNotification('app.registration.detailsSaveSuccessMessage');
-              resolve(true);
-            },
-            () => {
-              reject(true);
-            },
-          );
+          if (file.status !== 'error') {
+            deleteFile(
+              file.url.split('files/')[1].split('/')[0],
+              data => {
+                successNotification('app.registration.detailsSaveSuccessMessage');
+                resolve(true);
+              },
+              () => {
+                reject(true);
+              },
+            );
+          } else {
+            resolve(true);
+          }
         },
         onCancel: () => {
           reject(true);
